inference/requirements.txt
absl-py
affine
aiobotocore
aiohttp
aioitertools
anyio
astunparse
attrs
boto3
botocore
cartopy
cffi
cfgrib
charset-normalizer
click
click-plugins
cligj
earthaccess
eccodes
einops
filelock
flatbuffers
flexcache
flexparser
fonttools
frozendict
fsspec
gast
goes2go
google-pasta
grpcio
h5netcdf==1.8.1
h5py==3.14.0
herbie-data
httpcore
httpx
idna
jinja2
jmespath
keras
libclang
markdown
markupsafe
metpy
ml-dtypes==0.3.2
multidict
namex
netCDF4==1.6.5
networkx
numpy==1.26.4
opencv-python-headless
opt-einsum
optree
pandas
pillow
pint
pooch
pqdm
progressbar2
protobuf
pycparser
pydot
pyhdf
pyparsing
pyproj
python-dotenv
pyyaml
rasterio
requests
rich
rioxarray
s3fs
scikit-learn
scipy
shapely
tqdm
matplotlib
tensorflow
xarray
cftime
joblib
bottleneck
numexpr